Finished

+++ src/app/pages/books/books.component.ts
@@ -4,6 +4,7 @@ import { BookService } from '../../services/book.service';
 import { AuthorModel } from '../../models/author.model';
 import { AuthorService } from '../../services/author/author.service';
 import { map } from 'rxjs/operators';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-books',
@@ -33,7 +34,27 @@ export class BooksComponent implements OnInit {
 }
 
   deleteBook(idAuthor,idBook,i){
-    this.bookService.deleteBook(idAuthor, idBook);
+    Swal.fire({
+      title: 'Esta seguro que desea eliminar el libro?',
+      text: "No vas a ser capaz de revertir este cambio!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminalo!'
+    }).then((result) => {
+      if (result.value) {
+        this.bookList.splice(i, 1);
+        this.bookService.deleteBook(idAuthor, idBook);
+        Swal.fire(
+          'Libro eliminado!',
+          'El libro ha sido eliminado.',
+          'success'
+        )
+      }
+    })
+
+
 
   }
